+++ CHANGELOG.md
@@ -9,6 +9,8 @@ project adheres to
 
 ## Unreleased
 
+* Make the `calc(...)` function signal an error when args are known to
+  be invalid css (PR #138).
 * Minor fix in whitespace around at-rules.
 * Fix a typo in LICENSE (Issue #136).
 * Update sass-spec test suite to 2022-04-07.